Consolidate some code in the Provider, create new helper functions
